Add templating mechanism to generate .c/.h files
Generate the c/h code for constants definitions and objects initialization.
The generated files are generated under `src/` and `include/` as to not
conflict with the files from `ext/rbs_extension` and will soon replace
`constants.{h,c}` and `ruby_objs.{h,c}`.

We make sure the Rake task builds the required templates so we can compile
the extension.

Signed-off-by: Alexandre Terrasa <alexandre.terrasa@shopify.com>