Handle rbs_node and rbs_node_list types
This adds support for `rbs_node` and `rbs_node_list` field types in the
bindings generator. This enables access to child nodes and node lists
from parent AST nodes.

We handle fields named `type` (which is a Rust keyword) by escaping the
accessor method to `r#type()` and correctly mapping to the underlying
`type_` field in the C struct.